Menambahkan schema favourite untuk memberikan favorit ke jenis kucing yang diinginkan
