+++ src/database.js
@@ -13,21 +13,5 @@ db.once('open', () => {
     console.log('Succesfully Connected to Database');
 });
 
-//menyatakan schema database 
-const catSchema = mongoose.Schema({
-    "id" : Number, 
-    "breed" : String, 
-    "origin" : String, 
-    "country_code" : String, 
-    "life_span" : String, 
-    "characteristic" : String, 
-    "description"  : String,
-    "image_url" : String, 
-    "width" : Number, "height" : Number, 
-    "mime_type" : String,
-});
-
-//mengekspor model dari schema yang sudah dibuat
-module.exports = mongoose.model('Cat', catSchema);
